requirements.txt
streamlit==1.22.0
pandas==1.5.3
Pillow==9.5.0
python-dateutil==2.8.2 